Use real generated labs client; hand-author only the iam list binding
Cloned astronomer/astro to check the actual, already-merged spec for
astronomer/astro#39781 instead of guessing:

- Verified the labs bulk create/delete endpoints match what was
  hand-written (path, 1000-item cap on *both* create and delete,
  response shapes), added "AllowedIpAddressRange" to the labs
  -include-tags in the Makefile, and regenerated
  internal/clients/labs/api.gen.go for real. The hand-written labs
  client file is no longer needed and is deleted.
- Discovered the labs API has no list/read endpoint at all - only
  bulk create/delete. Reads actually belong to the pre-existing iam
  v1beta1 endpoint (GET /organizations/{id}/allowed-ip-address-ranges,
  tag "AllowedIpAddressRange"), so the resource now lists through the
  already-vendored iam client instead of a fabricated labs list call.
- Adding "AllowedIpAddressRange" to the iam client's -include-tags
  regenerates cleanly on its own, but oapi-codegen v2.1.0 renames the
  unrelated ListUsersParamsSorts constants in the process (a
  constant-naming collision with the new
  ListAllowedIpAddressRangesParamsSorts's overlapping enum values,
  which would break data_source_users_list.go - unrelated to this PR.
  Rather than carry that fix here, internal/clients/iam/api.gen.go is
  left untouched and only the list binding is hand-authored in
  internal/clients/iam/allowed_ip_address_ranges.go, using the exact
  types/shapes confirmed from the real spec and controller code.

Updated the resource to list via the iam client (with real
pagination via totalCount) and create/delete via the labs client.

EOF
)
